Don't call script automatically, only rely on target output
